pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lane-ai"
version = "0.1.0"
description = "Lane line detection using classical CV with an extensible DL stub"
authors = [{name = "Mobin Yousefi", email = ""}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
  "opencv-python>=4.8",
  "numpy>=1.24",
  "scipy>=1.11",
  "matplotlib>=3.8",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "ruff>=0.6.2", "black>=24.8"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
select = ["E","F","I","B"]

[tool.black]
line-length = 100